Point update
I made the puck go faster whenever it hits the paddle, which works out
quite well. I think that it could be improved to base the pucks speed on
the mouse’s speed when the puck hits the paddle. I also extended your
player’s point system to include losses and wins.