Created the one and onely Pong game.
I created a functional left paddle, created the basic correct ball
functionality, expanded the point system, and expanded on the GUI to
show the expanded point system. I also deleted a bunch of not needed
project files. We still need to improve the ball movements, fix the odd
bug, improve on the point system, work on the GUI (including the paused
state), and call it done!!